Arreglando el estropicio

+++ src/index.php
@@ -12,27 +12,61 @@ use Phroute\Phroute\RouteCollector;
 // Instancia del objeto router
 $router = new RouteCollector();
 
-// Rutas de la aplicación
-$router->get('/', function(){
-	return 'Estoy en la pagina principal';
+
+//Definir los filtros de las rutas
+$router->filter('auth',function(){
+
+	if(!isset($_SESSION['user'])) {
+		header('Location: /login');
+		return false;
+	}
 });
 
+
+$router->filter('admin',function(){
+
+	if (isset($_SESSION['user']) && !$_SESSION['user']->isAdmin()){
+		header('Location: /error');
+		return false;
+	}
+});
+
+$router->get('/error',function(){
+	if (isset($_SESSION['error'])){
+		$error = $_SESSION['error'];
+	}else{
+		$error = "Error desconocido";
+	}
+	include_once DIRECTORIO_VISTAS_FRONTEND."404.php";
+});
+
+//Definir las rutas de mi aplicación
+
+$router->get('/',function(){
+	return include_once DIRECTORIO_VISTAS_FRONTEND."indice.php";
+});
+
+$router->get('/admin',function(){
+	return include_once DIRECTORIO_VISTAS_BACKEND."indexAdmin.php";
+});
+
+
 //Rutas de Usuario CRUD
-//Rutas asociadas a la vistas de usuarios
-$router->get('/user/{id}/edit', [UserController::class, 'edit']);
-$router->get('/user/create', [UserController::class, 'create']);
+//Rutas asociadas a las vistas de usuario
+$router->get('/user/{id}/edit',[UserController::class,'edit'],["before"=>'auth']);
+$router->get('/user/create',[UserController::class,'create']);
 $router->get('/login',[UserController::class,'show_login']);
 $router->post('/user/login',[UserController::class,'verify']);
-$router->get('/logout',[UserController::class,'logout']);
+$router->get('/user/logout',[UserController::class,'logout'],["before"=>'auth']);
 
-//Rutas para la aplicación web visual
-$router->get('/user',[UserController::class,'index']);
-$router->get('/user/{id}',[UserController::class,'show']);
+//Rutas para la aplicacion web visual
+$router->get('/user',[UserController::class,'index'],["before"=>'admin']);
+$router->get('/user/{id}',[UserController::class,'show'],["before"=>'auth']);
 $router->post('/user',[UserController::class,'store']);
 $router->put('/user/{id}',[UserController::class,'update']);
-$router->delete('/user/{id}',[UserController::class,'destroy']);
+$router->delete('/user/{id}',[UserController::class,'destroy'],["before"=>'admin']);
 
-//Rutas de Servvicio API REST
+//Rutas de Servicio API REST
 $router->get('/api/user',[UserController::class,'index']);
 $router->get('/api/user/{id}',[UserController::class,'show']);
 $router->post('/api/user',[UserController::class,'store']);
@@ -45,7 +79,7 @@ $router->delete('/api/user/{id}',[UserController::class,'destroy']);
 
 
 $router->get('/control', function(){
-	include_once DIRECTORIO_VISTAS_BACKEND . "welcome.php";
+	include_once DIRECTORIO_VISTAS_BACKEND . "indexAdmin.php";
 });
